Add files via upload
Version 2 - archive size halved

+++ rANS_Coder/Program.cs
@@ -11,13 +11,12 @@ namespace rANS_Coder
 
             string fileName; // имя файла
             byte pow;  // степень размера для таблицы кодирования
-            
+
 
             if (args.Length == 0) // нет параметров командной строки
             {
                 throw new ArgumentNullException("argument out");
             }
-            //Console.WriteLine(args.Length);
 
 
 
@@ -41,8 +40,6 @@ namespace rANS_Coder
                 }
             }
 
-            //Console.WriteLine(" pow " + pow + " fileName " + fileName);
-
             if (pow < 8 || pow > 30) throw new ArgumentException("pow should be > 8 and pow < 30"); // проверка выхода pow за пределы
 
 
@@ -50,14 +47,14 @@ namespace rANS_Coder
 
 
             byte[] bcache = breader.ReadFile(fileName); // весь файл сохраняем в массиве байт
-            //Console.WriteLine("--- " + breader.FlagRNS);
+                                                        //Console.WriteLine("--- " + breader.FlagRNS);
 
 
             if (!breader.FlagRNS)
             { // кодирование
                 BinWriter bw = new BinWriter(); // подготовка записи закодированного файла
                 byte[] ext = breader.GetExtension(); // берем расширение для сохранения внутри файла
-                //Console.WriteLine(ext.Length);
+                                                     //Console.WriteLine(ext.Length);
                 Array.Reverse(bcache); // для кодирование разворачиваем исходный массив байт
                 bw.WriteFile(fileName, (byte)ext.Length); // 1 - записываем размер расширения
                 bw.AppToFile(ext); // 2 - записываем расширение
@@ -66,7 +63,7 @@ namespace rANS_Coder
                 Encoder enc = new Encoder(bcache, pow, 77777); // подготовка кодирования массива bcache со степенью pow с начальной инициализацией кодировщика (77777) - также проверяется на это число в раскодировке
 
                 byte[] TableFtoS = enc.CalcTabFtoS(); // рассчет таблицы для вычисления байт при дальнейшей раскодировке
-                //Console.WriteLine("TableFtoS " + TableFtoS.Length);
+                                                      //Console.WriteLine("TableFtoS " + TableFtoS.Length);
 
 
 
@@ -74,68 +71,52 @@ namespace rANS_Coder
 
 
 
-                UInt32[] xArray = enc.GetX(); // получаем массив чисел - состояние кодировщика
+                UInt16[] xArray = enc.GetX(); // получаем массив чисел - состояние кодировщика
+
 
                 bw.AppToFile(xArray); // 5 - записываем состояние кодировщика
-
-                //Console.WriteLine("--------------------------");
-                //for (int i = 0; i < xArray.Length; i++)
-                //{
-                //    Console.WriteLine(i + " " + xArray[i]);
-                //}
 
 
 
             }
             else // декодирование
             {
-                
+
                 byte[] ext = new byte[bcache[0]]; // читаем исходное расширение
                 for (int i = 0; i < ext.Length; i++)
                 {
                     ext[i] = bcache[i + 1];
                 }
                 string ExtensionOrigFile = Encoding.ASCII.GetString(ext);
-                //Console.WriteLine(ext.Length + " > " + ExtensionOrigFile);
                 UInt32 powInd = (UInt32)ext.Length + 1;  // место pow в запакованном файле
                 pow = bcache[powInd]; // читаем POW
 
                 UInt32 SizeTable = 1u << pow;
 
-                //Console.WriteLine("pow is " + pow + " size Table is " + (SizeTable));
 
                 byte[] TableFtoS = new byte[SizeTable]; // читаем таблицу для декодирования размером 2 в степени pow
                 Array.Copy(bcache, powInd + 1, TableFtoS, 0, SizeTable);
-                //Console.WriteLine("Length TableFtoS " + TableFtoS.Length);
 
 
                 UInt32 PackTableInd = powInd + 1 + SizeTable; // место начала закодированных данных в запакованном файле
                 UInt32 PackTableSize = (UInt32)bcache.Length - PackTableInd;
-                //Console.WriteLine("PackTableSize " + PackTableSize);
                 byte[] PackTable = new byte[PackTableSize]; // читаем упакованные данные для распаковки и записи в дальнейшем в распакованный файл
 
                 Array.Copy(bcache, PackTableInd, PackTable, 0, PackTableSize);
 
-                //for (int i = 0; i < PackTable.Length; i++)
-                //{
-                //    Console.WriteLine(i + " -- " + PackTable[i]);
-                //}
-
                 Decoder dec = new Decoder(TableFtoS, PackTable, pow); // готовим декодирование PackTable
 
-                byte [] OUT = dec.GetS(); // раскодированнй массив
+                byte[] OUT = dec.GetS(); // раскодированнй массив
 
                 BinWriter bw = new BinWriter(); // подготовка записи раскодированного файла
                 bw.WriteFile(fileName, ExtensionOrigFile, OUT);
 
-
-
             }
 
 
-
-
         }
+
+
 
     }
 }
